move back KafkaTopicSpec
